# api/openapi.yaml
openapi: 3.1.0
info:
  title: tripmap agent API
  version: 0.4.2
  description: |
    Authenticated itinerary API (MCP tools + optional REST). Schema v2: places
    catalog plus day route/stops as place refs. Notes are human-authored —
    put enrichment in places.id.info, not notes (usage rule, not enforced).
    Use update_day to change a day's title/notes/hike/ferry when the user asks.
    Human viewers sign in with Hellō, then use /me/trips/{id}/ (no capability tokens).
servers:
  - url: "{{BASE_URL}}"
paths:
...
  /api/agent/schema:
    get:
      operationId: getSchema
      summary: Itinerary schema and version
      x-audiences: [mcp]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Schema
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchemaInfo"
        "401":
          description: Unauthorized
  /api/agent/trips:
    get:
      operationId: listTrips
      summary: List itinerary IDs
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Trip ID list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripList"
        "401":
          description: Unauthorized
    post:
      operationId: createTrip
      summary: Create itinerary
      security:
        - bearerAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTripRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MutateResult"
        "401":
          description: Unauthorized
  /api/agent/trips/{id}:
    get:
      operationId: getTrip
      summary: Compact trip title card (not full routes)
      description: |
        Returns id, title, description, start, day count — not per-day routes.
        For overnight/endpoint edits use getTripYAML with scope=full (source of
        truth) then replaceDayRoutes. Prefer getTripYAML over getTrip when you
        need stops; chat defaults to day-neighborhood scope.
      x-audiences: [mcp]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripSummary"
        "404":
          description: Not found
    patch:
      operationId: patchTrip
      summary: Patch places info, day narrative, or structure
      description: Prefer over putTripYAML. Use update_day for title/notes/hike/ferry. Enrich via places.id.info. Also upsert_stop, remove_stop, swap_days (reorder whole days only — never overnight/endpoint), insert_day, delete_day. For overnight/endpoint or full route replacement, use replaceDayRoutes (not upsert_stop or swap_days). Do not put enrichment in notes.
      x-audiences: [mcp]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripPatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MutateResult"
  /api/agent/trips/{id}/yaml:
    get:
      operationId: getTripYAML
      summary: Get itinerary YAML
      description: |
        Source of truth for routes/stops before edits.
        Pass scope=day to get days day-1..day+1 plus only places referenced by
        those days. Pass scope=full (or omit scope on HTTP GET) for the entire
        itinerary.
      x-audiences: [mcp]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: scope
          in: query
          required: false
          description: |
            day = neighborhood YAML (days N±1 + referenced places).
            full = entire itinerary.
          schema:
            type: string
            enum: [day, full]
        - name: day
          in: query
          required: false
          description: Center day for scope=day.
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: YAML document
          content:
            text/plain:
              schema:
                type: string
    put:
      operationId: putTripYAML
      summary: Replace YAML (raw text body)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          text/plain:
            schema:
              type: string
              description: Full itinerary YAML
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MutateResult"
  /api/agent/trips/{id}/versions:
    get:
      operationId: listVersions
      summary: List YAML versions
      x-audiences: [mcp]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Versions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VersionList"
  /api/agent/trips/{id}/versions/{version_id}:
    get:
      operationId: getVersion
      summary: Get YAML for a prior version
      description: |
        Read-only fetch of a historical revision. Call listVersions first.
        Prefer this over restoreVersion when inspecting history.
      x-audiences: [mcp]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: version_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: YAML document for that version
          content:
            text/plain:
              schema:
                type: string
        "404":
          description: Not found
  /api/agent/trips/{id}/restore:
    post:
      operationId: restoreVersion
      summary: Restore a prior YAML version
      description: |
        Restores TO the given version_id (makes that revision current). To undo
        only the latest change, pass the previous version_id from listVersions
        (first entry with is_latest false) — not the current/latest id.
      x-audiences: [mcp]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestoreRequest"
      responses:
        "200":
          description: Restored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MutateResult"
  /api/agent/trips/{id}/day-routes:
    post:
      operationId: replaceDayRoutes
      summary: Replace full day routes (overnight / endpoint changes)
      description: |
        Replace the full route array for one or more days in a single call.
        Use this for overnight or day-endpoint changes — include both day N and
        day N+1 when continuity matters, with updated titles. For day N+1, keep
        existing mid/end stops; only change the start place — do not collapse a
        travel day to a single stop. Set each route[].type to the day role:
        lodging ends overnight (not via/depart), road waypoints via, sights
        attraction/etc. Optional places creates/updates the catalog in the same
        request. Do not use upsert_stop or swap_days for this.
      x-audiences: [mcp]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplaceDayRoutesRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MutateResult"
components:
  schemas:
...
    SchemaInfo:
      type: object
      properties:
        schema_version:
          type: integer
        description:
          type: string
        notes_policy:
          type: string
          description: Human-authored notes usage rule for agents (not enforced)
        fields:
          type: object
          additionalProperties:
            type: string
        patch_ops:
          type: array
          items:
            type: string
        days_patch_example:
          type: object
          additionalProperties: true
    TripList:
      type: object
      properties:
        trips:
          type: array
          items:
            type: string
    CreateTripRequest:
      type: object
      required:
        - id
        - yaml
      properties:
        id:
          type: string
        yaml:
          type: string
    TripSummary:
      type: object
      properties:
        id:
          type: string
        version_id:
          type: string
        schema_version:
          type: integer
        trip:
          type: string
        description:
          type: string
        start:
          type: string
        days:
          type: integer
    TripPatch:
      type: object
      additionalProperties: false
      description: |
        One or more structured ops. Prefer update_day for day narrative and
        places.<id>.info for enrichment. Day calendar dates are derived from
        trip.start + (day − 1) when start is set — do not patch per-day dates.
      properties:
        swap_days:
          type: array
          description: >
            Exactly two day numbers whose entire day bodies are exchanged (reorder).
            Do not use for changing an overnight or day endpoint — that needs route
            edits on the affected days (and the next day's start), not a swap.
          items:
            type: integer
          minItems: 2
          maxItems: 2
        update_day:
          $ref: "#/components/schemas/UpdateDay"
        places:
          type: object
          description: >
            Map of kebab-case place id to partial place fields. For a new place,
            title, lat, lon, and type are required at the top level of that object.
            Never put title under info — info is structured enrichment only
            (see PlaceInfo) and is deep-merged.
          additionalProperties:
            $ref: "#/components/schemas/PlacePatch"
        days:
          type: object
          description: >
            Day-number string keys (e.g. "10") to partial day updates. Prefer
            update_day for title/notes alone. To change an overnight or reorder
            a route — especially when the next day's start must change too —
            prefer replaceDayRoutes. Include title (and notes if the narrative
            should change) on those days.
          additionalProperties:
            $ref: "#/components/schemas/DayPatch"
        upsert_stop:
          $ref: "#/components/schemas/UpsertStop"
        remove_stop:
          $ref: "#/components/schemas/RemoveStop"
        delete_day:
          type: integer
        insert_day:
          $ref: "#/components/schemas/InsertDay"
...
    MutateResult:
      type: object
      properties:
        id:
          type: string
        version_id:
          type: string
        schema_version:
          type: integer
        viewer_url:
          type: string
          description: Signed-in tripmap viewer path /me/trips/{id}/ (Hellō login required)
        bundle_ok:
          type: boolean
        bundle_error:
          type: string
    VersionList:
      type: object
      properties:
        id:
          type: string
        versions:
          type: array
          items:
            type: object
            properties:
              version_id:
                type: string
              last_modified:
                type: string
              is_latest:
                type: boolean
    RestoreRequest:
      type: object
      required:
        - version_id
      properties:
        version_id:
          type: string
    ReplaceDayRoutesRequest:
      type: object
      additionalProperties: false
      required:
        - days
      properties:
        places:
          type: object
          description: >
            Optional place catalog upserts (kebab-case ids). New places need
            top-level title, lat, lon, and type — not under info.
          additionalProperties:
            $ref: "#/components/schemas/PlacePatch"
        days:
          type: object
          description: >
            Day-number string keys (e.g. "7"). Each entry replaces that day's
            full route. Include title when the overnight/endpoint changes.
            Typically update day N and day N+1 together. On each route, mark
            start/end lodging as overnight so the viewer can label Depart.
          minProperties: 1
          additionalProperties:
            type: object
            additionalProperties: false
            required:
              - route
            properties:
              title:
                type: string
              notes:
                type: string
              route:
                type: array
                minItems: 1
                items:
                  $ref: "#/components/schemas/StopRef"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer